breaking apart cytoscape layout into smaller helpers
