+++ client/routes/workflow/components/workflow-graph/helpers/cytoscape-layout.js
@@ -19,143 +19,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import { orderBy } from 'lodash-es';
+import arrangeGraph from './arrange-graph';
 import { CYTOSCAPE_LAYOUT_DEFAULTS, CYTOSCAPE_LAYOUT_NAME } from '../constants';
-
-const makeKey = (primary, secondary) => `${primary}-${secondary}`;
-
-/**
- * Traverse the graph recursively and set `level` and `parentTimeIndexOffset` for all nodes
- * @method arrangeNodes
- * @param {object} options
- * @param {object} options.idToChildren - node id to child lookup hash
- * @param {object} options.level - horizontal offset of the tree-like graph node arranged to avoid overlapping.
- * @param {object} options.nodes
- * @param {object} options.parentTimeIndex
- * @param {object} options.parentTimeIndexOffset - secondary time coordinate:
- *    there are two connected nodes: A -----> B
- *    A.data.timeIndex === B.data.timeIndex
- *  then we set
- *    A.data.timeIndexSecondary = 0
- *    B.data.timeIndexSecondary = 1
- *  To display B node beflow the A node whilst they they have the same timestamp
- */
-const arrangeNodes = ({
-  idToChildren,
-  level = 0,
-  nodes,
-  parentTimeIndex = -1,
-  parentTimeIndexOffset = 0,
-}) => {
-  let levelOffset = level;
-
-  nodes.forEach((node, index) => {
-    levelOffset += index ? 1 : 0;
-    const children = idToChildren[node.id()];
-    const nodeScratch = node.scratch(CYTOSCAPE_LAYOUT_NAME);
-
-    nodeScratch.level = levelOffset;
-    nodeScratch.timeIndexSecondary =
-      parentTimeIndex === nodeScratch.timeIndex ? parentTimeIndexOffset + 1 : 0;
-
-    if (children.length) {
-      const { level: newLevel } = arrangeNodes({
-        idToChildren,
-        level: levelOffset,
-        nodes: children,
-        parentTimeIndex: nodeScratch.timeIndex,
-        parentTimeIndexOffset: nodeScratch.timeIndexSecondary,
-      });
-
-      levelOffset = newLevel;
-    }
-  });
-
-  return {
-    level: levelOffset,
-  };
-};
-
-const arrangeGraph = ({ nodes, edges }, options) => {
-  const idToNode = {};
-  const idToChildren = {};
-  const timeIndices = {};
-  const hasParent = {};
-  const rootNodes = [];
-
-  // Iterate through nodes and set `timeIndex` and `timeIndexSecondary` for every node
-  // timeIndex is an index of the node timestamp in a sorted array of all timestamps
-  // Which is used to keep the chronogical order of nodes but discard the scale of time intervals
-  // when drawing the graph
-  nodes.forEach(node => {
-    idToNode[node.id()] = node;
-    idToChildren[node.id()] = [];
-    const timestamp = node.data().timestamp;
-
-    if (timeIndices[timestamp] === undefined) {
-      timeIndices[timestamp] = Object.keys(timeIndices).length;
-    }
-
-    node.scratch(CYTOSCAPE_LAYOUT_NAME, {
-      timeIndex: timeIndices[timestamp],
-      timeIndexSecondary: 0,
-    });
-  });
-
-  // Find all roots (entry points with no parents) of the graph
-  for (const edge of edges) {
-    idToChildren[edge.data().source].push(idToNode[edge.data().target]);
-    hasParent[edge.data().target] = true;
-  }
-
-  for (const node of nodes) {
-    if (!hasParent[node.id()]) {
-      rootNodes.push(node);
-    }
-  }
-
-  // Arrange all nodes in the graph, starting from root entry points
-  arrangeNodes({ idToChildren, nodes: rootNodes });
-
-  // Calculate `tTimes` for all (timeIndex, timeIndexSecondary) pairs
-  // which are the time (Y) coordinates nodes in the graph
-  let t = 0;
-  let prevT1 = 0;
-  let prevT2 = 0;
-  const tTimes = {};
-  const times = orderBy(
-    nodes.map(n => ({
-      t1: n.scratch(CYTOSCAPE_LAYOUT_NAME).timeIndex,
-      t2: n.scratch(CYTOSCAPE_LAYOUT_NAME).timeIndexSecondary,
-    })),
-    ['t1', 't2']
-  );
-
-  times.forEach(({ t1, t2 }) => {
-    const key = makeKey(t1, t2);
-
-    if (t1 === prevT1) {
-      t += (t2 - prevT2) * options.secondaryTimeStep;
-    }
-
-    t += (t1 - prevT1) * options.timeStep;
-    prevT1 = t1;
-    prevT2 = t2;
-
-    tTimes[key] = t;
-  });
-
-  // Set the `position` for all nodes using the calculated `level` and `tTimes` values
-  nodes.forEach(node => {
-    const nodeScratch = node.scratch(CYTOSCAPE_LAYOUT_NAME);
-    const key = makeKey(nodeScratch.timeIndex, nodeScratch.timeIndexSecondary);
-
-    nodeScratch.position = {
-      x: nodeScratch.level * options.levelStep,
-      y: tTimes[key],
-    };
-  });
-};
 
 class CytoscapeLayout {
   constructor(options) {
